Oh so many changes! Most of them quite useful...

+++ blocks/block_1in1out.cpp
@@ -74,12 +74,12 @@ void Block_1in1out::process()
 	v_in = get_data_ptr< integer_t >(sig_in1_);
 	v_out = get_data_ptr< integer_t >(sig_out1_);
 
-	for(uint16_t i=0; i<sig_out1_->get_frame_size(); i++)
+	
+	for(uint32_t i=0; i<sig_out1_->get_frame_size(); i++)
 	{
 		v_out[i] = 1+v_in[i];
 	}
 	
-//  	usleep(100000);
 	sig_out1_->send();
 }
